+++ services/horizon/internal/ingest/processors/asset_stats_processor.go
@@ -2,7 +2,6 @@ package processors
 
 import (
 	"database/sql"
-	"math/big"
 
 	"github.com/stellar/go/ingest"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
@@ -109,8 +108,10 @@ func (p *AssetStatsProcessor) Commit() error {
 	assetStatsDeltas := p.assetStatSet.All()
 	for _, delta := range assetStatsDeltas {
 		var rowsAffected int64
+		var stat history.ExpAssetStat
+		var err error
 
-		stat, err := p.assetStatsQ.GetAssetStat(
+		stat, err = p.assetStatsQ.GetAssetStat(
 			delta.AssetType,
 			delta.AssetCode,
 			delta.AssetIssuer,
@@ -121,39 +122,53 @@ func (p *AssetStatsProcessor) Commit() error {
 		}
 
 		if assetStatNotFound {
-			// Insert
-			if delta.NumAccounts < 0 {
+			// Safety checks
+			if delta.Accounts.Authorized < 0 {
 				return ingest.NewStateError(errors.Errorf(
-					"NumAccounts negative but DB entry does not exist for asset: %s %s %s",
+					"Authorized accounts negative but DB entry does not exist for asset: %s %s %s",
+					delta.AssetType,
+					delta.AssetCode,
+					delta.AssetIssuer,
+				))
+			} else if delta.Accounts.AuthorizedToMaintainLiabilities < 0 {
+				return ingest.NewStateError(errors.Errorf(
+					"AuthorizedToMaintainLiabilities accounts negative but DB entry does not exist for asset: %s %s %s",
+					delta.AssetType,
+					delta.AssetCode,
+					delta.AssetIssuer,
+				))
+			} else if delta.Accounts.Unauthorized < 0 {
+				return ingest.NewStateError(errors.Errorf(
+					"Unauthorized accounts negative but DB entry does not exist for asset: %s %s %s",
 					delta.AssetType,
 					delta.AssetCode,
 					delta.AssetIssuer,
 				))
 			}
 
+			// Insert
 			var errInsert error
 			rowsAffected, errInsert = p.assetStatsQ.InsertAssetStat(delta)
 			if errInsert != nil {
 				return errors.Wrap(errInsert, "could not insert asset stat")
 			}
 		} else {
-			statBalance, ok := new(big.Int).SetString(stat.Amount, 10)
-			if !ok {
-				return errors.New("Error parsing: " + stat.Amount)
+			var statBalances assetStatBalances
+			if err = statBalances.Parse(&stat.Balances); err != nil {
+				return errors.Wrap(err, "Error parsing balances")
 			}
 
-			deltaBalance, ok := new(big.Int).SetString(delta.Amount, 10)
-			if !ok {
-				return errors.New("Error parsing: " + stat.Amount)
+			var deltaBalances assetStatBalances
+			if err = deltaBalances.Parse(&delta.Balances); err != nil {
+				return errors.Wrap(err, "Error parsing balances")
 			}
 
-			// statBalance = statBalance + deltaBalance
-			statBalance.Add(statBalance, deltaBalance)
-			statAccounts := stat.NumAccounts + delta.NumAccounts
+			statBalances = statBalances.Add(deltaBalances)
+			statAccounts := stat.Accounts.Add(delta.Accounts)
 
-			if statAccounts == 0 {
+			if statAccounts.IsZero() {
 				// Remove stats
-				if statBalance.Cmp(big.NewInt(0)) != 0 {
+				if !statBalances.IsZero() {
 					return ingest.NewStateError(errors.Errorf(
 						"Removing asset stat by final amount non-zero for: %s %s %s",
 						delta.AssetType,
@@ -175,8 +190,10 @@ func (p *AssetStatsProcessor) Commit() error {
 					AssetType:   delta.AssetType,
 					AssetCode:   delta.AssetCode,
 					AssetIssuer: delta.AssetIssuer,
-					Amount:      statBalance.String(),
-					NumAccounts: statAccounts,
+					Accounts:    statAccounts,
+					Balances:    statBalances.ConvertToHistoryObject(),
+					Amount:      statBalances.Authorized.String(),
+					NumAccounts: statAccounts.Authorized,
 				})
 				if err != nil {
 					return errors.Wrap(err, "could not update asset stat")
@@ -202,50 +219,26 @@ func (p *AssetStatsProcessor) adjustAssetStat(
 	preTrustline *xdr.TrustLineEntry,
 	postTrustline *xdr.TrustLineEntry,
 ) error {
-	var deltaBalance xdr.Int64
-	var deltaAccounts int32
-	var trustline xdr.TrustLineEntry
+	deltaAccounts := map[xdr.Uint32]int32{}
+	deltaBalances := map[xdr.Uint32]int64{}
 
-	if preTrustline != nil && postTrustline == nil {
-		trustline = *preTrustline
-		// removing a trustline
-		if xdr.TrustLineFlags(preTrustline.Flags).IsAuthorized() {
-			deltaAccounts = -1
-			deltaBalance = -preTrustline.Balance
-		}
-	} else if preTrustline == nil && postTrustline != nil {
-		trustline = *postTrustline
-		// adding a trustline
-		if xdr.TrustLineFlags(postTrustline.Flags).IsAuthorized() {
-			deltaAccounts = 1
-			deltaBalance = postTrustline.Balance
-		}
-	} else if preTrustline != nil && postTrustline != nil {
-		trustline = *postTrustline
-		// updating a trustline
-		if xdr.TrustLineFlags(preTrustline.Flags).IsAuthorized() &&
-			xdr.TrustLineFlags(postTrustline.Flags).IsAuthorized() {
-			// trustline remains authorized
-			deltaAccounts = 0
-			deltaBalance = postTrustline.Balance - preTrustline.Balance
-		} else if xdr.TrustLineFlags(preTrustline.Flags).IsAuthorized() &&
-			!xdr.TrustLineFlags(postTrustline.Flags).IsAuthorized() {
-			// trustline was authorized and became unauthorized
-			deltaAccounts = -1
-			deltaBalance = -preTrustline.Balance
-		} else if !xdr.TrustLineFlags(preTrustline.Flags).IsAuthorized() &&
-			xdr.TrustLineFlags(postTrustline.Flags).IsAuthorized() {
-			// trustline was unauthorized and became authorized
-			deltaAccounts = 1
-			deltaBalance = postTrustline.Balance
-		}
-		// else, trustline was unauthorized and remains unauthorized
-		// so there is no change to accounts or balances
-	} else {
+	if preTrustline == nil && postTrustline == nil {
 		return ingest.NewStateError(errors.New("both pre and post trustlines cannot be nil"))
 	}
 
-	err := p.assetStatSet.AddDelta(trustline.Asset, int64(deltaBalance), deltaAccounts)
+	var trustline xdr.TrustLineEntry
+	if preTrustline != nil {
+		trustline = *preTrustline
+		deltaAccounts[preTrustline.Flags] -= 1
+		deltaBalances[preTrustline.Flags] -= int64(preTrustline.Balance)
+	}
+	if postTrustline != nil {
+		trustline = *postTrustline
+		deltaAccounts[postTrustline.Flags] += 1
+		deltaBalances[postTrustline.Flags] += int64(postTrustline.Balance)
+	}
+
+	err := p.assetStatSet.AddDelta(trustline.Asset, deltaBalances, deltaAccounts)
 	if err != nil {
 		return errors.Wrap(err, "error running AssetStatSet.AddDelta")
 	}