+++ services/horizon/internal/ingest/processors/asset_stats_processor_test.go
@@ -58,6 +58,12 @@ func (s *AssetStatsProcessorTestSuiteState) TestCreateTrustLine() {
 			AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
 			AssetIssuer: trustLineIssuer.Address(),
 			AssetCode:   "EUR",
+			Accounts:    history.ExpAssetStatAccounts{Authorized: 1},
+			Balances: history.ExpAssetStatBalances{
+				Authorized:                      "0",
+				AuthorizedToMaintainLiabilities: "0",
+				Unauthorized:                    "0",
+			},
 			Amount:      "0",
 			NumAccounts: 1,
 		},
@@ -83,8 +89,21 @@ func (s *AssetStatsProcessorTestSuiteState) TestCreateTrustLineUnauthorized() {
 	})
 	s.Assert().NoError(err)
 
-	s.mockQ.
-		On("InsertAssetStats", []history.ExpAssetStat{}, maxBatchSize).Return(nil).Once()
+	s.mockQ.On("InsertAssetStats", []history.ExpAssetStat{
+		{
+			AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
+			AssetIssuer: trustLineIssuer.Address(),
+			AssetCode:   "EUR",
+			Accounts:    history.ExpAssetStatAccounts{Unauthorized: 1},
+			Balances: history.ExpAssetStatBalances{
+				Authorized:                      "0",
+				AuthorizedToMaintainLiabilities: "0",
+				Unauthorized:                    "0",
+			},
+			Amount:      "0",
+			NumAccounts: 0,
+		},
+	}, maxBatchSize).Return(nil).Once()
 }
 
 func TestAssetStatsProcessorTestSuiteLedger(t *testing.T) {
@@ -223,8 +242,37 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestInsertTrustLine() {
 		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
 		AssetIssuer: trustLineIssuer.Address(),
 		AssetCode:   "EUR",
+		Accounts: history.ExpAssetStatAccounts{
+			Authorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "10",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
 		Amount:      "10",
 		NumAccounts: 1,
+	}).Return(int64(1), nil).Once()
+
+	s.mockQ.On("GetAssetStat",
+		xdr.AssetTypeAssetTypeCreditAlphanum4,
+		"USD",
+		trustLineIssuer.Address(),
+	).Return(history.ExpAssetStat{}, sql.ErrNoRows).Once()
+	s.mockQ.On("InsertAssetStat", history.ExpAssetStat{
+		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
+		AssetIssuer: trustLineIssuer.Address(),
+		AssetCode:   "USD",
+		Accounts: history.ExpAssetStatAccounts{
+			Unauthorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "0",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "10",
+		},
+		Amount:      "0",
+		NumAccounts: 0,
 	}).Return(int64(1), nil).Once()
 
 	s.Assert().NoError(s.processor.Commit())
@@ -273,6 +321,12 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestUpdateTrustLine() {
 		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
 		AssetIssuer: trustLineIssuer.Address(),
 		AssetCode:   "EUR",
+		Accounts:    history.ExpAssetStatAccounts{Authorized: 1},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "100",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
 		Amount:      "100",
 		NumAccounts: 1,
 	}, nil).Once()
@@ -280,6 +334,12 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestUpdateTrustLine() {
 		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
 		AssetIssuer: trustLineIssuer.Address(),
 		AssetCode:   "EUR",
+		Accounts:    history.ExpAssetStatAccounts{Authorized: 1},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "110",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
 		Amount:      "110",
 		NumAccounts: 1,
 	}).Return(int64(1), nil).Once()
@@ -290,28 +350,44 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestUpdateTrustLine() {
 func (s *AssetStatsProcessorTestSuiteLedger) TestUpdateTrustLineAuthorization() {
 	lastModifiedLedgerSeq := xdr.Uint32(1234)
 
-	trustLine := xdr.TrustLineEntry{
+	// EUR trustline: 100 unauthorized -> 10 authorized
+	eurTrustLine := xdr.TrustLineEntry{
 		AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
 		Asset:     xdr.MustNewCreditAsset("EUR", trustLineIssuer.Address()),
 		Balance:   100,
 	}
-	updatedTrustLine := xdr.TrustLineEntry{
+	eurUpdatedTrustLine := xdr.TrustLineEntry{
 		AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
 		Asset:     xdr.MustNewCreditAsset("EUR", trustLineIssuer.Address()),
 		Balance:   10,
 		Flags:     xdr.Uint32(xdr.TrustLineFlagsAuthorizedFlag),
 	}
 
-	otherTrustLine := xdr.TrustLineEntry{
+	// USD trustline: 100 authorized -> 10 unauthorized
+	usdTrustLine := xdr.TrustLineEntry{
 		AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
 		Asset:     xdr.MustNewCreditAsset("USD", trustLineIssuer.Address()),
 		Balance:   100,
 		Flags:     xdr.Uint32(xdr.TrustLineFlagsAuthorizedFlag),
 	}
-	otherUpdatedTrustLine := xdr.TrustLineEntry{
+	usdUpdatedTrustLine := xdr.TrustLineEntry{
 		AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
 		Asset:     xdr.MustNewCreditAsset("USD", trustLineIssuer.Address()),
 		Balance:   10,
+	}
+
+	// ETH trustline: 100 authorized -> 10 authorized_to_maintain_liabilities
+	ethTrustLine := xdr.TrustLineEntry{
+		AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
+		Asset:     xdr.MustNewCreditAsset("ETH", trustLineIssuer.Address()),
+		Balance:   100,
+		Flags:     xdr.Uint32(xdr.TrustLineFlagsAuthorizedFlag),
+	}
+	ethUpdatedTrustLine := xdr.TrustLineEntry{
+		AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
+		Asset:     xdr.MustNewCreditAsset("ETH", trustLineIssuer.Address()),
+		Balance:   10,
+		Flags:     xdr.Uint32(xdr.TrustLineFlagsAuthorizedToMaintainLiabilitiesFlag),
 	}
 
 	err := s.processor.ProcessChange(ingest.Change{
@@ -320,14 +396,14 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestUpdateTrustLineAuthorization() 
 			LastModifiedLedgerSeq: lastModifiedLedgerSeq - 1,
 			Data: xdr.LedgerEntryData{
 				Type:      xdr.LedgerEntryTypeTrustline,
-				TrustLine: &trustLine,
+				TrustLine: &eurTrustLine,
 			},
 		},
 		Post: &xdr.LedgerEntry{
 			LastModifiedLedgerSeq: lastModifiedLedgerSeq,
 			Data: xdr.LedgerEntryData{
 				Type:      xdr.LedgerEntryTypeTrustline,
-				TrustLine: &updatedTrustLine,
+				TrustLine: &eurUpdatedTrustLine,
 			},
 		},
 	})
@@ -339,14 +415,33 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestUpdateTrustLineAuthorization() 
 			LastModifiedLedgerSeq: lastModifiedLedgerSeq - 1,
 			Data: xdr.LedgerEntryData{
 				Type:      xdr.LedgerEntryTypeTrustline,
-				TrustLine: &otherTrustLine,
+				TrustLine: &usdTrustLine,
 			},
 		},
 		Post: &xdr.LedgerEntry{
 			LastModifiedLedgerSeq: lastModifiedLedgerSeq,
 			Data: xdr.LedgerEntryData{
 				Type:      xdr.LedgerEntryTypeTrustline,
-				TrustLine: &otherUpdatedTrustLine,
+				TrustLine: &usdUpdatedTrustLine,
+			},
+		},
+	})
+	s.Assert().NoError(err)
+
+	err = s.processor.ProcessChange(ingest.Change{
+		Type: xdr.LedgerEntryTypeTrustline,
+		Pre: &xdr.LedgerEntry{
+			LastModifiedLedgerSeq: lastModifiedLedgerSeq - 1,
+			Data: xdr.LedgerEntryData{
+				Type:      xdr.LedgerEntryTypeTrustline,
+				TrustLine: &ethTrustLine,
+			},
+		},
+		Post: &xdr.LedgerEntry{
+			LastModifiedLedgerSeq: lastModifiedLedgerSeq,
+			Data: xdr.LedgerEntryData{
+				Type:      xdr.LedgerEntryTypeTrustline,
+				TrustLine: &ethUpdatedTrustLine,
 			},
 		},
 	})
@@ -356,11 +451,33 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestUpdateTrustLineAuthorization() 
 		xdr.AssetTypeAssetTypeCreditAlphanum4,
 		"EUR",
 		trustLineIssuer.Address(),
-	).Return(history.ExpAssetStat{}, sql.ErrNoRows).Once()
-	s.mockQ.On("InsertAssetStat", history.ExpAssetStat{
+	).Return(history.ExpAssetStat{
 		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
 		AssetIssuer: trustLineIssuer.Address(),
 		AssetCode:   "EUR",
+		Accounts: history.ExpAssetStatAccounts{
+			Unauthorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "0",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "100",
+		},
+		Amount:      "0",
+		NumAccounts: 0,
+	}, nil).Once()
+	s.mockQ.On("UpdateAssetStat", history.ExpAssetStat{
+		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
+		AssetIssuer: trustLineIssuer.Address(),
+		AssetCode:   "EUR",
+		Accounts: history.ExpAssetStatAccounts{
+			Authorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "10",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
 		Amount:      "10",
 		NumAccounts: 1,
 	}).Return(int64(1), nil).Once()
@@ -373,18 +490,78 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestUpdateTrustLineAuthorization() 
 		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
 		AssetIssuer: trustLineIssuer.Address(),
 		AssetCode:   "USD",
+		Accounts: history.ExpAssetStatAccounts{
+			Authorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "100",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
 		Amount:      "100",
 		NumAccounts: 1,
 	}, nil).Once()
-	s.mockQ.On("RemoveAssetStat",
+	s.mockQ.On("UpdateAssetStat", history.ExpAssetStat{
+		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
+		AssetIssuer: trustLineIssuer.Address(),
+		AssetCode:   "USD",
+		Accounts: history.ExpAssetStatAccounts{
+			Unauthorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "0",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "10",
+		},
+		Amount:      "0",
+		NumAccounts: 0,
+	}).Return(int64(1), nil).Once()
+
+	s.mockQ.On("GetAssetStat",
 		xdr.AssetTypeAssetTypeCreditAlphanum4,
-		"USD",
+		"ETH",
 		trustLineIssuer.Address(),
-	).Return(int64(1), nil).Once()
+	).Return(history.ExpAssetStat{
+		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
+		AssetIssuer: trustLineIssuer.Address(),
+		AssetCode:   "ETH",
+		Accounts: history.ExpAssetStatAccounts{
+			Authorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "100",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
+		Amount:      "100",
+		NumAccounts: 1,
+	}, nil).Once()
+	s.mockQ.On("UpdateAssetStat", history.ExpAssetStat{
+		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
+		AssetIssuer: trustLineIssuer.Address(),
+		AssetCode:   "ETH",
+		Accounts: history.ExpAssetStatAccounts{
+			AuthorizedToMaintainLiabilities: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "0",
+			AuthorizedToMaintainLiabilities: "10",
+			Unauthorized:                    "0",
+		},
+		Amount:      "0",
+		NumAccounts: 0,
+	}).Return(int64(1), nil).Once()
+
 	s.Assert().NoError(s.processor.Commit())
 }
 
 func (s *AssetStatsProcessorTestSuiteLedger) TestRemoveTrustLine() {
+	authorizedTrustLine := xdr.TrustLineEntry{
+		AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
+		Asset:     xdr.MustNewCreditAsset("EUR", trustLineIssuer.Address()),
+		Balance:   0,
+		Flags:     xdr.Uint32(xdr.TrustLineFlagsAuthorizedFlag),
+	}
 	unauthorizedTrustLine := xdr.TrustLineEntry{
 		AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
 		Asset:     xdr.MustNewCreditAsset("USD", trustLineIssuer.Address()),
@@ -395,13 +572,8 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestRemoveTrustLine() {
 		Type: xdr.LedgerEntryTypeTrustline,
 		Pre: &xdr.LedgerEntry{
 			Data: xdr.LedgerEntryData{
-				Type: xdr.LedgerEntryTypeTrustline,
-				TrustLine: &xdr.TrustLineEntry{
-					AccountId: xdr.MustAddress("GAOQJGUAB7NI7K7I62ORBXMN3J4SSWQUQ7FOEPSDJ322W2HMCNWPHXFB"),
-					Asset:     xdr.MustNewCreditAsset("EUR", trustLineIssuer.Address()),
-					Balance:   0,
-					Flags:     xdr.Uint32(xdr.TrustLineFlagsAuthorizedFlag),
-				},
+				Type:      xdr.LedgerEntryTypeTrustline,
+				TrustLine: &authorizedTrustLine,
 			},
 		},
 		Post: nil,
@@ -428,6 +600,14 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestRemoveTrustLine() {
 		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
 		AssetIssuer: trustLineIssuer.Address(),
 		AssetCode:   "EUR",
+		Accounts: history.ExpAssetStatAccounts{
+			Authorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "0",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
 		Amount:      "0",
 		NumAccounts: 1,
 	}, nil).Once()
@@ -436,6 +616,32 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestRemoveTrustLine() {
 		"EUR",
 		trustLineIssuer.Address(),
 	).Return(int64(1), nil).Once()
+
+	s.mockQ.On("GetAssetStat",
+		xdr.AssetTypeAssetTypeCreditAlphanum4,
+		"USD",
+		trustLineIssuer.Address(),
+	).Return(history.ExpAssetStat{
+		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
+		AssetIssuer: trustLineIssuer.Address(),
+		AssetCode:   "USD",
+		Accounts: history.ExpAssetStatAccounts{
+			Unauthorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "0",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
+		Amount:      "0",
+		NumAccounts: 0,
+	}, nil).Once()
+	s.mockQ.On("RemoveAssetStat",
+		xdr.AssetTypeAssetTypeCreditAlphanum4,
+		"USD",
+		trustLineIssuer.Address(),
+	).Return(int64(1), nil).Once()
+
 	s.Assert().NoError(s.processor.Commit())
 }
 
@@ -496,6 +702,14 @@ func (s *AssetStatsProcessorTestSuiteLedger) TestProcessUpgradeChange() {
 		AssetType:   xdr.AssetTypeAssetTypeCreditAlphanum4,
 		AssetIssuer: trustLineIssuer.Address(),
 		AssetCode:   "EUR",
+		Accounts: history.ExpAssetStatAccounts{
+			Authorized: 1,
+		},
+		Balances: history.ExpAssetStatBalances{
+			Authorized:                      "10",
+			AuthorizedToMaintainLiabilities: "0",
+			Unauthorized:                    "0",
+		},
 		Amount:      "10",
 		NumAccounts: 1,
 	}).Return(int64(1), nil).Once()