+++ services/horizon/internal/resourceadapter/asset_stat.go
@@ -23,9 +23,14 @@ func PopulateAssetStat(
 	res.Asset.Type = xdr.AssetTypeToString[row.AssetType]
 	res.Asset.Code = row.AssetCode
 	res.Asset.Issuer = row.AssetIssuer
+	res.Accounts = protocol.AssetStatAccounts(row.Accounts)
 	res.Amount, err = amount.IntStringToAmount(row.Amount)
 	if err != nil {
 		return errors.Wrap(err, "Invalid amount in PopulateAssetStat")
+	}
+	err = populateAssetStatBalances(&res.Balances, row.Balances)
+	if err != nil {
+		return err
 	}
 	res.NumAccounts = row.NumAccounts
 	flags := int8(issuer.Flags)
@@ -44,4 +49,23 @@ func PopulateAssetStat(
 	}
 	res.Links.Toml = hal.NewLink(toml)
 	return
+}
+
+func populateAssetStatBalances(res *protocol.AssetStatBalances, row history.ExpAssetStatBalances) (err error) {
+	res.Authorized, err = amount.IntStringToAmount(row.Authorized)
+	if err != nil {
+		return errors.Wrapf(err, "Invalid amount in PopulateAssetStatBalances: %q", row.Authorized)
+	}
+
+	res.AuthorizedToMaintainLiabilities, err = amount.IntStringToAmount(row.AuthorizedToMaintainLiabilities)
+	if err != nil {
+		return errors.Wrapf(err, "Invalid amount in PopulateAssetStatBalances: %q", row.AuthorizedToMaintainLiabilities)
+	}
+
+	res.Unauthorized, err = amount.IntStringToAmount(row.Unauthorized)
+	if err != nil {
+		return errors.Wrapf(err, "Invalid amount in PopulateAssetStatBalances: %q", row.Unauthorized)
+	}
+
+	return nil
 }